==> SpatialDataAnalysis_withR/01_vector.Rmd <==
---
title: "Vector data in R"
output: 
  html_document: 
    theme: yeti
---

```{r, echo=FALSE, message=FALSE, results='hide', purl=FALSE}
## This chunk automatically generates a text .R version of this script when running within knitr.  You do not need to run this...
input  = knitr::current_input()  # filename of input document
output = paste(tools::file_path_sans_ext(input), 'R', sep = '.')
knitr::purl(input,output,documentation=2,quiet=T,envir=new.env())
knitr::opts_chunk$set(cache=TRUE)
```

[<i class="fa fa-file-code-o fa-3x" aria-hidden="true"></i> The R Script associated with this page is available here](`r output`).  Download this file and open it (or copy-paste into a new script) with RStudio if you want to follow along.  


# Working with vector data

Before running any code here, make sure that you have installed all [necessary packages](packages.html) beforehand. 

As described in the [lecture](lecture.html), spatial vector data come in a variety of formats and there are also multiple ways of loading them into R.

## Vector data in text format

Sometimes spatial data is distributed in the form on text information only. A good example are for instance point localities. 
For start, lets get some spatial data in text format

```{r}
# Lets get some data from
#Fritz, Steffen; See, Linda; Perger, Christoph; McCallum, Ian; Schill, Christian; Schepaschenko, Dmitry; Duerauer, Martina; Karner, Mathias; Dresel, Christopher; Laso-Bayas, Juan-Carlos; Lesiv, Myroslava; Moorthy, Inian; Salk, Carl F; Danylo, Olha; Sturn, Tobias; Albrecht, Franziska; You, Liangzhi; Kraxner, Florian; Obersteiner, Michael (2016): A global dataset of crowdsourced land cover and land use reference data (2011-2012). PANGAEA https://doi.pangaea.de/10.1594/PANGAEA.869660

dat <- read.table('https://doi.pangaea.de/10.1594/PANGAEA.869660?format=textfile',
                  skip = 17,sep = '\t',header = T)

class(dat)
```
As you can see, this data is currently in tabular format. Let's make it explicitly spatial. 
To do so we are going to work with the **sp** package for start. 

The **sp** package is the old default way to create and manipulate spatial objects in R. Many R packages, particular those that haven't been updated in a while, will likely still depend on this format. However it has largely been superseded by the *sf* package that combines spatial information with the tidy data concept. More about this latter though.

Convert and format our tabular data to a spatial object

```{r}
library(sp)

# Set the coordinates for this object to the coordinate columns
# Here we first define the x-axis (longitude), followed by the y-axis (latitude)
coordinates(dat) <- ~ Longitude + Latitude

# By default this object won't have any projection assigned
# Since we know that the tabular data is in WGS-84 projection (Longitude-Latitude), we can set it
proj4string(dat) <- '+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'

```

In case you are not aware of the projections string in proj4 format, go to this [site](https://spatialreference.org/) to find a comprehensive list.

<div class="well">
Do you think you can visualize this new object and check that it is indeed spatial?


<button data-toggle="collapse" class="btn btn-primary btn-sm round" data-target="#demo1">Show Solution</button>
<div id="demo1" class="collapse">
```{r, purl=F}
plot(dat)

class(dat)
```
Well done. 
</div>
</div>

You note that the object has now been converted to a SpatialPointsDataFrame, so a Spatial Points object, which is a requirement for many packages that interact with spatial data

### Well known text information

Text format can be used not only to define point data (such as longitude-latitude), but also other geometric representations such as polygon data. According OGC standards these data can be for instance formatted as [Well-known text (WKT)](https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry), which is an important concept in spatial analyis that we will mention latter too.

For instance here is polygon that covers the Laxenburg schlosspark

```{r}
lax <- 'POLYGON ((16.3594574385736 48.0702355060673,16.3609882350041 48.0694683197585,16.3648152260803 48.0688107223911,16.3684782032532 48.0682627181663,16.3736173055555 48.0696509842027,16.3794124634709 48.0710757446162,16.3814353016112 48.0709296171584,16.380724574697 48.0694683197585,16.3785923939546 48.0664725303671,16.3749840880827 48.0617958276381,16.3690249162641 48.0575206573412,16.3621910036281 48.0518198780156,16.3587467116595 48.0578129879854,16.3537716232604 48.0564609448419,16.3553024196909 48.0589822939637,16.3534435954539 48.0606631247802,16.3532249102496 48.0634400293739,16.3544823501746 48.0651207146451,16.3591840820682 48.0663263898385,16.3598948089823 48.0668013450392,16.3583640125519 48.067605105394,16.3590200681649 48.0688472557986,16.3594574385736 48.0702355060673))'

```

To convert it to an sp object, we will rely on the GEOS library available through R with the *rgeos* package

```{r}
library(rgeos)
laxpol <- readWKT(lax)

```

### Creating your own spatial data

We won't go much into detail in this and in my experience you will rarely do this directly in R. Here is for instance how we could define a Spatial Points object directly in R. If you are more interested in this, check out some of the [resources](resources.html).

```{r}

# First lets create some random point data
coords <- data.frame( x = rnorm(100,3),
                      y = rnorm(100,3)
                      )

# To convert to a SpatialPoints object
x = SpatialPoints(coords)

# Finally lets assign some data to it, so that it becomes a Spatial*DataFrame object
d <- data.frame(ID = 1:100)
spdf <- SpatialPointsDataFrame(coords,d)

```


<div class="well">
This data still doesn't have a projection. Remember how to assign it?

<button data-toggle="collapse" class="btn btn-primary btn-sm round" data-target="#demo2">Show Solution</button>
<div id="demo2" class="collapse">
```{r, purl=F}
proj4string(spdf) <- '+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'
```

</div>
</div>

# Loading and writing different formats of vector data

In most cases however the spatial data won't come in text format, but in some other format that requires separate solutions for loading them in. Here we going to use the **rgdal** package which supports a number of spatial data and is the main background work hose also for [raster](02_raster.html) data.

```{r, echo = F}
library(rgdal)

knitr::kable(ogrDrivers())
```

## Read in some external data

For testing we are going to read in some of the example data that comes with the rgdal package

```{r}
# Data from the rgdal package
dsn <- system.file("vectors", package = "rgdal")[1]

# Which layers are in the data source
ogrListLayers(dsn)

```


The rgdal package always requires a folder name (data source) and a filename without extension.

<div class="well">
Can you use the **ogrInfo()** command from the rgdal package to query the information in the 'cities' layer?
The command requires a folder (dsn) and the name of the layer

<button data-toggle="collapse" class="btn btn-primary btn-sm round" data-target="#demo3">Show Solution</button>
<div id="demo3" class="collapse">
```{r, purl=F}
ogrInfo(dsn=dsn, layer="cities")
```

</div>
</div>

To load the layer using rgdal

```{r}
# To load the cities.shp file. Note that you point to the name of the layer and not the name of the file
cities <- readOGR(dsn=dsn, layer="cities")

```

## Writing vector data

The rgdal package also comes with functionalities to write vector data

```{r, eval=FALSE}
# Save the loaded cities file to your folder
output_folder = '.' # a dot is for the current folder
output_name = 'new_cities' # the output name
writeOGR(cities,dsn = output_folder, layer = output_name ,driver = "ESRI Shapefile")
```

<div class="well">
Can you load and write the previously created laxenburg park data used above (as WKT) and save it (any format you like)?

<button data-toggle="collapse" class="btn btn-primary btn-sm round" data-target="#demo4">Show Solution</button>
<div id="demo4" class="collapse">

```{r, purl=FALSE, eval = FALSE}
library(rgeos)
library(rgdal)

lax <- 'POLYGON ((16.3594574385736 48.0702355060673,16.3609882350041 48.0694683197585,16.3648152260803 48.0688107223911,16.3684782032532 48.0682627181663,16.3736173055555 48.0696509842027,16.3794124634709 48.0710757446162,16.3814353016112 48.0709296171584,16.380724574697 48.0694683197585,16.3785923939546 48.0664725303671,16.3749840880827 48.0617958276381,16.3690249162641 48.0575206573412,16.3621910036281 48.0518198780156,16.3587467116595 48.0578129879854,16.3537716232604 48.0564609448419,16.3553024196909 48.0589822939637,16.3534435954539 48.0606631247802,16.3532249102496 48.0634400293739,16.3544823501746 48.0651207146451,16.3591840820682 48.0663263898385,16.3598948089823 48.0668013450392,16.3583640125519 48.067605105394,16.3590200681649 48.0688472557986,16.3594574385736 48.0702355060673))'

laxpol <- readWKT(lax)

# Convert to data.frame
laxpol <- SpatialPolygonsDataFrame(laxpol, data = data.frame(ID = 1, name = 'Laxenburg park') )
proj4string(laxpol) <- '+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'

# Write the output
writeOGR(laxpol,dsn = '.',layer = 'Laxenburg',driver = "ESRI Shapefile")
# or as GPKG, using the database as folder
writeOGR(laxpol,dsn = 'Laxenburg.gpkg',layer = 'Laxenburg',driver = "GPKG")

```
Well done!
</div>
</div>

# Vector data with the sf package

The R programming language has in recent years experience some form of revolution that data are treated and formatted as ['tidy' data](https://vita.had.co.nz/papers/tidy-data.html). 

Loading spatial data into R using **sf** is comparably easier

```{r}
library(sf)

# Load the default country file coming with sf
nc <- st_read(system.file("gpkg/nc.gpkg", package="sf"))

class(nc)

# Show the top 6 rows
head(nc, 6)

```

Notice how this new object is both a 'sf' object as well as a 'data.frame'. All the data assigned to this vector table can be used like you would use any other spatial information in a R data.frame, e.g. subsetting, filtering, plotting, etc...
Furthermore the geometry is here not saved as part of the R-object but explicitly adressable in the column 'geom' using WKT. 

Similarly, individual (or multiple) can be plotted by adressing them as you would adress an item in a list object
So

```{r}
plot(nc['SID74'])
```

Writing outputs in sf is equally straight forward.

```{r, eval = FALSE}
# If you want to write this into a folder called myfolder
st_write(nc, paste0('myfolder', "/", "nc.shp") )

# You can also output the geometry data in WKT format
st_as_text(nc$geom)

```

Sometimes a function or r-package creates outputs that are not in sf format. These need to be converted to sf before being able to use them in tidy format.
The function for that in R is `st_as_sf`.

<div class="well">
Can you take the Laxenburg park object and calculate its area ?
The function to do so is 'st_area()' and requires a sf object.

<button data-toggle="collapse" class="btn btn-primary btn-sm round" data-target="#demo5">Show Solution</button>
<div id="demo5" class="collapse">

```{r, purl=FALSE, eval = FALSE}
# Convert the file
sf_laxpol <- st_as_sf(laxpol)

# The area in m2
st_area(sf_laxpol)

# Similarly one can write the laxenburg data to geopackage easily with sf
st_write(new_sf, 'Laxenburg.gpkg')

```
Area calculations using the sf package use the 'units' package to assign a unit in R to vector of numeric values.
You should get in result of ~2.2 Million ha
</div>
</div>

For the rest of the course we will primarily use the **sf** package for any spatial manipulation.
